Refactor hauki errors and request to own python file
TILA-801
